requirements.txt
antlr4-python3-runtime==4.7.1
asn1crypto==0.24.0
async-timeout==3.0.0; python_version >= '3.5.3'
attrs==18.2.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
idna==2.7
itsdangerous==0.24
markupsafe==1.0
mongoengine==0.15.3
multidict==4.4.2; python_version >= '3.4.1'
multipledispatch==0.6.0
pycparser==2.19; python_version != '3.1.*'
pydatalog==0.17.1
pymongo==3.7.1
requests==2.19.1
six==1.11.0
toolz==0.9.0
ujson==1.35
unification==0.2.2
urllib3==1.23; python_version != '3.3.*'
yarl==1.2.6; python_version >= '3.4.1'
pytest >= "4.4.0"
pytest-aiohttp==0.3.0
